+++ mnmp-admin/src/main/resources/application-local.yml
@@ -11,4 +11,6 @@ spring:
     hikari:
       minimum-idle: 1
 excel:
-  path: "F:/excel/"
+  path:
+    task: "F:/excel/"
+    white-list:
